Less dependencies, multi-touch supported
I rewrote/refactored a large portion of the code. Here's the summary:

* Removed `d3.js` dependency. We don't need it, we can use pure vanilla JS.
* Removed another unused dependency.
* Changed the demo page to use JavaScript modules.
* Rewrote a couple of the PathExtras functions to be shorter and simpler.
* Heavily refactored the main index.js logic. It is still a bit messy.
* We now have one "state" per active pointer. This allows multiple
  pointers (e.g. multiple fingers) simultaneously.
* Removed the interpolation logic. I don't think it is needed, but I may
  change my mind in the future.
* Used `getCoalescedEvents()` to fetch all events. This is a drawing
  app, we need to get all pointer events. See also:
  https://developer.chrome.com/blog/aligning-input-events/
* Dots can now be drawn. See also:
  https://github.com/Desousak/svg-pen-sketch/issues/15

The code is still a bit messy. This commit is a result of a night
without sleep. I'm happy with the user-facing result, I'm not entirely
happy with the code.

TODO: Multi-finger-touch isn't drawing smoothly for all fingers (on
Firefox Android). Why? This seems to work correctly.

TODO: We may try removing the `touchstart` and possibly the
`contextmenu` events.

TODO: Further rewrite the code to allow variable stroke-width by varying
the pressure.

TODO: We may add keyboard modifiers, such as Shift/Ctrl. But this
shouldn't be tightly coupled with the main library.

TODO: Further refactor the code to allow for multiple "tools" (i.e.
multiple actions or behaviors). Currently it only supports `drawing` and
`erasing`. It could/should support straight lines, poly-lines,
rectangles, circles, arrows, etc. These behaviors should be easy to
extend, and should be easy to remap on-the-fly.